requirements.txt
torch
torchvision
torchmetrics
pyyaml
tqdm
tensorboard
pillow
pandas
scikit-learn
matplotlib